Conflicts:
	src/plugins/clangcodemodel/clangdclient.cpp

+++ src/libs/utils/stringutils.cpp
@@ -268,7 +268,7 @@ QString asciify(const QString &input)
         if (c.isPrint() && c.unicode() < 128)
             result.append(c);
         else
-            result.append(QString::fromLatin1("u%1").arg(c.unicode(), 4, 16, QChar('0')));
+            result.append(QString::fromLatin1("u%1").arg(int16_t(c.unicode()), 4, 16, QChar('0')));
     }
     return result;
 }
